refacto: student registration to event
